Export usage event schema (#4122)
## Problem solved

Exports the Zod schema for usage events. This way it can be used for validate the message before sending it to the usage tracker queue.
Moved it under `core` because it is not tied to worker logic and can be used outside workers.

<!-- start pr-codex -->

---

## PR-Codex overview
The focus of this PR is to export `usageEventSchema` under `service-utils/core` for message validation before sending to the tracker.

### Detailed summary
- Exported `usageEventSchema` under `service-utils/core`
- Moved `UsageEvent` type declaration to `service-utils/core`
- Updated import in `cf-worker/usage.ts` to reflect changes

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
