+++ .changeset/nervous-suns-kick.md
@@ -1,0 +1,5 @@
+---
+"@thirdweb-dev/service-utils": patch
+---
+
+exported usageEventSchema under service-utils/core so it can be used to validate messages before sending to usage tracker
